Fix eval harness: per-class NMS, conf threshold 0.001, float boxes, 400 pre-NMS cap

+++ misc/model_output_handler.py
@@ -151,7 +151,7 @@ class Model_output_handler():
         scale_bboxes[:, 2] *= width
         scale_bboxes[:, 1] *= height
         scale_bboxes[:, 3] *= height
-        return (scale_bboxes.cpu().numpy()).astype(int)
+        return scale_bboxes.cpu().numpy()
 
 
 class UnNormalize(object):
